components/triage/requirements.txt
libclang
matplotlib
networkx
mypy-protobuf
pytest
python-dotenv
jsonschema
pika
sqlalchemy
psycopg2-binary
pyyaml
redis
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
mcp-agent
anthropic
uv
mcp-server-fetch
mcp-server-tree-sitter

./clusterfuzz/src/ ; python_version>='3.11'
pydantic==2.10.6
# NOTE: The following are extra requirements that aren't included if you just install
# clusterfuzz with its setup.py.
# Then, don't pin these: we'll use whatever versions pip decides is compatible with
# clusterfuzz during the install process.
# We don't really need these for our functionality in this little directory, but they do
# cause import errors if they're not included.
aiohttp
google-cloud-secret-manager
google-cloud-storage==1.44.0
google-cloud-core==1.7.3
google-cloud-datastore==1.12.0